Add automatic command redeployment
